refactor: minor refactor on servings tunnel
